TASK: Adjustments from review: Move namespace `ExceptionHandling` to `ErrorHandling`
